feat: 添加拖拽功能混入类，整合拖拽相关逻辑
